style changes and refactoring
