+++ src/sass/all.sass
@@ -14,6 +14,7 @@ $section-padding-medium: 5rem 1.5rem
 
 // Olipear colours
 $bg-light: #F8F8F8
+$bg-medium: #eaebf1
 $bg-dark: #0B2027
 $primary: #4CB497
 $secondary: #FA9F42
@@ -21,7 +22,7 @@ $text: #031926
 $shadow: #0b20274f
 
 $interact: #4E5166
-$hover: #333442
+$hover: #24252e
 $disabled: #B8B9C2
 
 //olipear styles
